Move all pipeline streams to separate modules
