+++ import.js
@@ -8,7 +8,7 @@ var peliasConfig = require( 'pelias-config' ).generate();
 var logger = require( 'pelias-logger' ).get( 'openaddresses' );
 
 var interpretUserArgs = require( './lib/interpretUserArgs' );
-var importPipelines = require( './lib/import_pipelines' );
+var importPipeline = require( './lib/importPipeline' );
 
 // Pretty-print the total time the import took.
 function startTiming() {
@@ -21,44 +21,6 @@ function startTiming() {
   });
 }
 
-/**
- * Import all OpenAddresses CSV files in a directory into Pelias elasticsearch.
- *
- * @param {array of string} files An array of the absolute file-paths to import.
- * @param {object} opts Options to configure the import. Supports the following
- *    keys:
- *
- *      deduplicate: Pass address object through `address-deduplicator-stream`
- *        to perform deduplication. See the documentation:
- *        https://github.com/pelias/address-deduplicator-stream
- *
- *      adminValues: Add admin values to each address object (since
- *        OpenAddresses doesn't contain any) using `admin-lookup`. See the
- *        documentation: https://github.com/pelias/admin-lookup
- */
-function importOpenAddressesFiles( files, opts ){
-  logger.info( 'Importing %s files.', files.length );
-  var pipeline = [ importPipelines.createFullRecordStream(files) ];
-
-
-  if( opts.deduplicate ){
-    pipeline.push(importPipelines.createDeduplicatorStream());
-  }
-
-  if( opts.adminValues ){
-    pipeline.push(importPipelines.createAdminLookupStream());
-  }
-
-  pipeline.push(importPipelines.createPeliasElasticsearchPipeline());
-
-  startTiming();
-
-  // start the import process by piping all the streams appropriately
-  for(var i = 0; i < pipeline.length - 1; i++) {
-    pipeline[i].pipe(pipeline[i+1]);
-  }
-}
-
 var args = interpretUserArgs.interpretUserArgs( process.argv.slice( 2 ) );
 
 if( 'exitCode' in args ){
@@ -66,7 +28,9 @@ if( 'exitCode' in args ){
   process.exit( args.exitCode );
 }
 else {
+  startTiming();
+
   var files = interpretUserArgs.getFileList(peliasConfig, args);
 
-  importOpenAddressesFiles( files, args );
+  importPipeline.create( files, args );
 }
