+++ lib/isValidCsvRecord.js
@@ -1,0 +1,10 @@
+/*
+ * Return true if a record has all of LON, LAT, NUMBER and STREET defined
+ */
+function isValidCsvRecord( record ){
+  return [ 'LON', 'LAT', 'NUMBER', 'STREET' ].every(function(prop) {
+    return record[ prop ].length > 0;
+  });
+}
+
+module.exports = isValidCsvRecord;
