+++ lib/streams/cleanupStream.js
@@ -1,6 +1,6 @@
 var through2 = require( 'through2' );
 
-var cleanup = require( './cleanup' );
+var cleanup = require( '../cleanup' );
 
 /*
  * create a stream that performs any needed cleanup on a record
@@ -23,5 +23,5 @@ function createCleanupStream() {
 }
 
 module.exports = {
-  createCleanupStream: createCleanupStream
+  create: createCleanupStream
 };
