+++ test/isValidCsvRecord.js
@@ -1,6 +1,6 @@
 var tape = require( 'tape' );
 
-var importPipelines = require( '../lib/import_pipelines' );
+var isValidCsvRecord = require( '../lib/isValidCsvRecord' );
 
 tape( 'Identifies invalid CSV records.', function ( test ){
   var records = [
@@ -9,10 +9,10 @@ tape( 'Identifies invalid CSV records.', function ( test ){
     {LON: '', LAT: '2', STREET: '', NUMBER: '4', SOME_PROP: 'value'}
   ];
   records.forEach( function ( rec ){
-    test.ok( !importPipelines.isValidCsvRecord( rec ), 'Record identified as invalid' );
+    test.ok( !isValidCsvRecord( rec ), 'Record identified as invalid' );
   });
 
   var validRecord = {LON: '1', LAT: '2', STREET: '3', NUMBER: '4', SOME_PROP: 'abs'};
-  test.ok( importPipelines.isValidCsvRecord( validRecord ), 'Record identified as valid.' );
+  test.ok( isValidCsvRecord( validRecord ), 'Record identified as valid.' );
   test.end();
 });
