+++ test/streams/documentStream.js
@@ -1,7 +1,7 @@
 var tape = require( 'tape' );
 var event_stream = require( 'event-stream' );
 
-var importPipelines = require( '../lib/import_pipelines' );
+var DocumentStream = require( '../../lib/streams/documentStream' );
 
 function test_stream(input, testedStream, callback) {
   var input_stream = event_stream.readArray(input);
@@ -15,14 +15,13 @@ tape( 'documentStream catches records with no street', function(test) {
     NUMBER: 5
   };
   var stats = { badRecordCount: 0 };
-  var documentStream = importPipelines.createDocumentStream(stats);
+  var documentStream = DocumentStream.create(stats);
 
   test_stream([input], documentStream, function(err, actual) {
     test.equal(actual.length, 0, 'no documents should be pushed' );
     test.equal(stats.badRecordCount, 1, 'bad record count updated');
     test.end();
   });
-
 });
 
 tape( 'documentStream does not set zipcode if zipcode is emptystring', function(test) {
@@ -34,7 +33,7 @@ tape( 'documentStream does not set zipcode if zipcode is emptystring', function(
     POSTCODE: ''
   };
   var stats = { badRecordCount: 0 };
-  var documentStream = importPipelines.createDocumentStream(stats);
+  var documentStream = DocumentStream.create(stats);
 
   test_stream([input], documentStream, function(err, actual) {
     test.equal(actual.length, 1, 'the document should be pushed' );
@@ -42,5 +41,4 @@ tape( 'documentStream does not set zipcode if zipcode is emptystring', function(
     test.equal(actual[0].getAddress('zip', undefined));
     test.end();
   });
-
 });
