+++ test/streams/elasticsearchStream.js
@@ -1,12 +1,12 @@
 var tape = require( 'tape' );
 
-var importPipelines = require( '../lib/import_pipelines' );
+var elasticsearchStream = require( '../../lib/streams/elasticsearchStream' );
 
 tape(
   'importPipeline.createPeliasElasticsearchPipeline() interface',
   function ( test ){
     test.plan( 1 );
-    var esPipeline = importPipelines.createPeliasElasticsearchPipeline();
+    var esPipeline = elasticsearchStream.create();
     test.ok( esPipeline.writable, 'Stream is writable' );
   }
 );
